Fix multiword search with incomplete words.
It would trip over if the dictionary contained
words with the same prefix starting with -.
Example is searching for
"sense des aiguilles" which would break the loop
upon encountering "-aige".
Which meant that searching for "sense des aig" would
find nothing. "sense des a" would however due to "a"
not being considered due to the stoplist.