Attempt to fix search sometimes not working.
ListView fails to process setSelection in some cases,
probably related to it being busy updating for a new
IndexAdapter.
Try to avoid this by posting the message to it instead
of to the uiHandler.
Might fix issue #120.
However the delay in jumpToTextFromHyperLink is still
needed, so this still is not really reliable.